Move SwerveModule to utils package
I will find a better place for it later.
